requirements.txt
pytest-playwright>=0.7.0
playwright>=1.35.0
beautifulsoup4
selenium>=4.21.0
webdriver-manager>=3.8.5

pandas

# data
python-dotenv>=0.21.0

